Merged "Mother" code and cleaned up license
The old repository (https://github.com/kalopa/mother) held the missing
files which used this gem. As part of a major reworking of the gem,
those executables are now included in here in the bin directory. I also
cleaned up the old copyright which was a mix of MIT and GPL in favour of
just GPL v2.0. This is the baseline check-in of the code. From here, the
daemons get cleaned up and tested.